10_IT_AI_Core_Layer/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-auth==2.23.3
google-cloud-bigquery==3.13.0
google-cloud-logging==3.8.0
google-generativeai>=0.8.0
google-api-python-client>=2.0.0
websockets>=12.0
numpy>=1.26.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
python-magic>=0.4.27
pymupdf>=1.23.0
pillow-heif>=0.13.0
pyyaml>=6.0.1
python-docx>=1.1.0
